# 118x faster than GCC 12: KL & JS divergence with AVX-512FP16