@ashvardanian 2023-10-23
:tada: This PR is included in version 3.2.0 :tada:

The release is available on [GitHub release](https://github.com/ashvardanian/SimSIMD/releases/tag/v3.2.0)

Your **[semantic-release](https://github.com/semantic-release/semantic-release)** bot :package::rocket: